agent-ingest/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
opentelemetry-proto==1.27.0
protobuf>=3.19.0,<5.0.0
